+++ src/main/java/jwp/controller/LogInController.java
@@ -1,0 +1,31 @@
+package jwp.controller;
+
+import core.db.MemoryUserRepository;
+import jwp.model.User;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import java.io.IOException;
+
+@WebServlet("/user/login")
+public class LogInController extends HttpServlet {
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        HttpSession session = req.getSession();
+        String userId = req.getParameter("userId");
+        String password = req.getParameter("password");
+        User logInUser = new User(userId, password);
+        User user = MemoryUserRepository.getInstance().findUserById(userId);
+
+        if (user != null && user.isSameUser(logInUser)) {
+            session.setAttribute("user", user);
+            resp.sendRedirect("/");
+            return;
+        }
+        resp.sendRedirect("/user/loginFailed.jsp");
+    }
+}
