+++ tomcat.8080/work/Tomcat/localhost/ROOT/org/apache/jsp/user/list_jsp.java
@@ -1,7 +1,7 @@
 /*
  * Generated by the Jasper component of Apache Tomcat
  * Version: Apache Tomcat/8.5.42
- * Generated at: 2023-03-22 10:42:22 UTC
+ * Generated at: 2023-03-24 06:48:08 UTC
  * Note: The last modified time of this file was set to
  *       the last modified time of the source file after
  *       generation to assist with modification tracking.
@@ -24,7 +24,7 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
   static {
     _jspx_dependants = new java.util.HashMap<java.lang.String,java.lang.Long>(4);
     _jspx_dependants.put("/include/header.jspf", Long.valueOf(1679479578255L));
-    _jspx_dependants.put("/include/navigation.jspf", Long.valueOf(1679479578255L));
+    _jspx_dependants.put("/include/navigation.jspf", Long.valueOf(1679640478993L));
     _jspx_dependants.put("jar:file:/Users/gimjeong-u/.gradle/caches/modules-2/files-2.1/javax.servlet/jstl/1.2/74aca283cd4f4b4f3e425f5820cda58f44409547/jstl-1.2.jar!/META-INF/c.tld", Long.valueOf(1153352682000L));
     _jspx_dependants.put("file:/Users/gimjeong-u/.gradle/caches/modules-2/files-2.1/javax.servlet/jstl/1.2/74aca283cd4f4b4f3e425f5820cda58f44409547/jstl-1.2.jar", Long.valueOf(1676908657118L));
   }
@@ -41,6 +41,9 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
     _jspx_imports_classes = null;
   }
 
+  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fchoose;
+  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest;
+  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fotherwise;
   private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems;
 
   private volatile javax.el.ExpressionFactory _el_expressionfactory;
@@ -81,10 +84,16 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
   }
 
   public void _jspInit() {
+    _005fjspx_005ftagPool_005fc_005fchoose = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
+    _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
+    _005fjspx_005ftagPool_005fc_005fotherwise = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
     _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
   }
 
   public void _jspDestroy() {
+    _005fjspx_005ftagPool_005fc_005fchoose.release();
+    _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.release();
+    _005fjspx_005ftagPool_005fc_005fotherwise.release();
     _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems.release();
   }
 
@@ -134,6 +143,8 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
       out.write("</head>");
       out.write("\n");
       out.write("<body>\n");
+      out.write("\n");
+      out.write("\n");
       out.write("<nav class=\"navbar navbar-expand-md navbar-dark bg-dark\" aria-label=\"Fourth navbar example\">\n");
       out.write("    <div class=\"container-fluid\">\n");
       out.write("        <a class=\"navbar-brand\" href=\"/\"> KUIT </a>\n");
@@ -171,8 +182,10 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
       out.write("        </ul>\n");
       out.write("\n");
       out.write("        <div class=\"col-md-3 text-end\">\n");
-      out.write("            <a href=\"/user/login.jsp\" type=\"button\" class=\"btn btn-outline-primary me-2\">Login</a>\n");
-      out.write("            <a href=\"/user/form.jsp\" type=\"button\" class=\"btn btn-primary\">Sign-up</a>\n");
+      out.write("            ");
+      if (_jspx_meth_c_005fchoose_005f0(_jspx_page_context))
+        return;
+      out.write("\n");
       out.write("        </div>\n");
       out.write("    </header>\n");
       out.write("</div>");
@@ -220,6 +233,115 @@ public final class list_jsp extends org.apache.jasper.runtime.HttpJspBase
     } finally {
       _jspxFactory.releasePageContext(_jspx_page_context);
     }
+  }
+
+  private boolean _jspx_meth_c_005fchoose_005f0(javax.servlet.jsp.PageContext _jspx_page_context)
+          throws java.lang.Throwable {
+    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
+    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
+    //  c:choose
+    org.apache.taglibs.standard.tag.common.core.ChooseTag _jspx_th_c_005fchoose_005f0 = (org.apache.taglibs.standard.tag.common.core.ChooseTag) _005fjspx_005ftagPool_005fc_005fchoose.get(org.apache.taglibs.standard.tag.common.core.ChooseTag.class);
+    boolean _jspx_th_c_005fchoose_005f0_reused = false;
+    try {
+      _jspx_th_c_005fchoose_005f0.setPageContext(_jspx_page_context);
+      _jspx_th_c_005fchoose_005f0.setParent(null);
+      int _jspx_eval_c_005fchoose_005f0 = _jspx_th_c_005fchoose_005f0.doStartTag();
+      if (_jspx_eval_c_005fchoose_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
+        do {
+          out.write("\n");
+          out.write("            ");
+          if (_jspx_meth_c_005fwhen_005f0(_jspx_th_c_005fchoose_005f0, _jspx_page_context))
+            return true;
+          out.write("\n");
+          out.write("            ");
+          if (_jspx_meth_c_005fotherwise_005f0(_jspx_th_c_005fchoose_005f0, _jspx_page_context))
+            return true;
+          out.write("\n");
+          out.write("            ");
+          int evalDoAfterBody = _jspx_th_c_005fchoose_005f0.doAfterBody();
+          if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
+            break;
+        } while (true);
+      }
+      if (_jspx_th_c_005fchoose_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
+        return true;
+      }
+      _005fjspx_005ftagPool_005fc_005fchoose.reuse(_jspx_th_c_005fchoose_005f0);
+      _jspx_th_c_005fchoose_005f0_reused = true;
+    } finally {
+      org.apache.jasper.runtime.JspRuntimeLibrary.releaseTag(_jspx_th_c_005fchoose_005f0, _jsp_getInstanceManager(), _jspx_th_c_005fchoose_005f0_reused);
+    }
+    return false;
+  }
+
+  private boolean _jspx_meth_c_005fwhen_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fchoose_005f0, javax.servlet.jsp.PageContext _jspx_page_context)
+          throws java.lang.Throwable {
+    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
+    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
+    //  c:when
+    org.apache.taglibs.standard.tag.rt.core.WhenTag _jspx_th_c_005fwhen_005f0 = (org.apache.taglibs.standard.tag.rt.core.WhenTag) _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.get(org.apache.taglibs.standard.tag.rt.core.WhenTag.class);
+    boolean _jspx_th_c_005fwhen_005f0_reused = false;
+    try {
+      _jspx_th_c_005fwhen_005f0.setPageContext(_jspx_page_context);
+      _jspx_th_c_005fwhen_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fchoose_005f0);
+      // /include/navigation.jspf(41,12) name = test type = boolean reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
+      _jspx_th_c_005fwhen_005f0.setTest(((java.lang.Boolean) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${not empty sessionScope.user}", boolean.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null)).booleanValue());
+      int _jspx_eval_c_005fwhen_005f0 = _jspx_th_c_005fwhen_005f0.doStartTag();
+      if (_jspx_eval_c_005fwhen_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
+        do {
+          out.write("\n");
+          out.write("                <a href=\"/user/logout\" role=\"button\" class=\"btn btn-outline-primary me-2\">로그아웃</a>\n");
+          out.write("                <a href=\"/user/updateForm?userId=");
+          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${sessionScope.user.userId}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+          out.write("\" role=\"button\" class=\"btn btn-primary\" >개인정보수정</a>\n");
+          out.write("            ");
+          int evalDoAfterBody = _jspx_th_c_005fwhen_005f0.doAfterBody();
+          if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
+            break;
+        } while (true);
+      }
+      if (_jspx_th_c_005fwhen_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
+        return true;
+      }
+      _005fjspx_005ftagPool_005fc_005fwhen_0026_005ftest.reuse(_jspx_th_c_005fwhen_005f0);
+      _jspx_th_c_005fwhen_005f0_reused = true;
+    } finally {
+      org.apache.jasper.runtime.JspRuntimeLibrary.releaseTag(_jspx_th_c_005fwhen_005f0, _jsp_getInstanceManager(), _jspx_th_c_005fwhen_005f0_reused);
+    }
+    return false;
+  }
+
+  private boolean _jspx_meth_c_005fotherwise_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fchoose_005f0, javax.servlet.jsp.PageContext _jspx_page_context)
+          throws java.lang.Throwable {
+    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
+    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
+    //  c:otherwise
+    org.apache.taglibs.standard.tag.common.core.OtherwiseTag _jspx_th_c_005fotherwise_005f0 = (org.apache.taglibs.standard.tag.common.core.OtherwiseTag) _005fjspx_005ftagPool_005fc_005fotherwise.get(org.apache.taglibs.standard.tag.common.core.OtherwiseTag.class);
+    boolean _jspx_th_c_005fotherwise_005f0_reused = false;
+    try {
+      _jspx_th_c_005fotherwise_005f0.setPageContext(_jspx_page_context);
+      _jspx_th_c_005fotherwise_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fchoose_005f0);
+      int _jspx_eval_c_005fotherwise_005f0 = _jspx_th_c_005fotherwise_005f0.doStartTag();
+      if (_jspx_eval_c_005fotherwise_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
+        do {
+          out.write("\n");
+          out.write("                <a href=\"/user/login.jsp\" type=\"button\" class=\"btn btn-outline-primary me-2\">Login</a>\n");
+          out.write("                <a href=\"/user/form.jsp\" type=\"button\" class=\"btn btn-primary\">Sign-up</a>\n");
+          out.write("            ");
+          int evalDoAfterBody = _jspx_th_c_005fotherwise_005f0.doAfterBody();
+          if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
+            break;
+        } while (true);
+      }
+      if (_jspx_th_c_005fotherwise_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
+        return true;
+      }
+      _005fjspx_005ftagPool_005fc_005fotherwise.reuse(_jspx_th_c_005fotherwise_005f0);
+      _jspx_th_c_005fotherwise_005f0_reused = true;
+    } finally {
+      org.apache.jasper.runtime.JspRuntimeLibrary.releaseTag(_jspx_th_c_005fotherwise_005f0, _jsp_getInstanceManager(), _jspx_th_c_005fotherwise_005f0_reused);
+    }
+    return false;
   }
 
   private boolean _jspx_meth_c_005fforEach_005f0(javax.servlet.jsp.PageContext _jspx_page_context)
